refactor(stores): extract dangling kind types into own files
AddRoomStore and RoomEditorStore each declared their *ErrorKind and
*SuccessKind sealed interfaces at file scope alongside the store contract.
Move each to its own file (same package, no consumer import changes) to
satisfy the single-type-per-file convention.
