+++ shared/feature-add-room/api/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/add_room/api/store/AddRoomStore.kt
@@ -66,16 +66,3 @@ interface AddRoomStore : Store<Intent, State, Label> {
         data class ShowError(val kind: AddRoomErrorKind) : Label
     }
 }
-
-sealed interface AddRoomErrorKind {
-    data object UnknownError : AddRoomErrorKind
-    data object SaveFailed : AddRoomErrorKind
-    data class RoomAlreadyExists(val roomNumber: Int, val floorNumber: Int) : AddRoomErrorKind
-    data object FloorNumberRequired : AddRoomErrorKind
-    data object RoomNumberRequired : AddRoomErrorKind
-    data object BedsCountRequired : AddRoomErrorKind
-}
-
-sealed interface AddRoomSuccessKind {
-    data class RoomCreated(val roomNumber: Int) : AddRoomSuccessKind
-}
