+++ shared/feature-add-room/api/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/add_room/api/store/AddRoomSuccessKind.kt
@@ -1,0 +1,5 @@
+package dev.nonoxy.residetrack.feature.add_room.api.store
+
+sealed interface AddRoomSuccessKind {
+    data class RoomCreated(val roomNumber: Int) : AddRoomSuccessKind
+}
