+++ shared/feature-room-editor/api/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/room_editor/api/store/RoomEditorSuccessKind.kt
@@ -1,0 +1,7 @@
+package dev.nonoxy.residetrack.feature.room_editor.api.store
+
+sealed interface RoomEditorSuccessKind {
+    data object StudentsSaved : RoomEditorSuccessKind
+    data object RoomUpdated : RoomEditorSuccessKind
+    data object RoomDeleted : RoomEditorSuccessKind
+}
